@@ -15,41 +15,24 @@
     "build": "npm run lint && node --experimental-json-modules esbuild.config.mjs production"
   },
   "devDependencies": {
-    "@skeletonlabs/tw-plugin": "^0.1.0",
-    "@sveltejs/adapter-auto": "^2.0.0",
+    "@biomejs/biome": "1.9.4",
     "@tsconfig/svelte": "^5.0.4",
     "@types/node": "^20.10.4",
-    "@typescript-eslint/eslint-plugin": "^5.45.0",
-    "@typescript-eslint/parser": "^5.45.0",
-    "autoprefixer": "^10.4.20",
     "builtin-modules": "^4.0.0",
-    "dotenv": "^16.4.5",
-    "esbuild": "^0.24.0",
-    "esbuild-svelte": "^0.8.2",
-    "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte": "^2.30.0",
     "moment": "^2.30.1",
-    "npm-run-all": "^4.1.5",
     "obsidian": "latest",
-    "postcss": "^8.4.45",
-    "postcss-cli": "^11.0.0",
-    "postcss-load-config": "^5.1.0",
-    "prettier": "^2.8.0",
-    "prettier-plugin-svelte": "^2.10.1",
-    "sass": "^1.69.5",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
-    "svelte-preprocess": "^5.1.1",
-    "tailwindcss": "^3.4.11",
+    "svelte": "^5.1.10",
+    "svelte-check": "^4.0.5",
+    "tailwindcss": "^3.4.14",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0"
+    "typescript": "^5.6.3"
   },
   "dependencies": {
     "@emoji-mart/data": "^1.2.1",
     "@floating-ui/dom": "^1.5.1",
     "clsx": "^2.0.0",
     "emoji-mart": "^5.6.0",
-    "svelte-writable-derived": "^3.1.1"
+    "svelte-writable-derived": "^3.1.1",
+    "tailwind-merge": "^2.5.4"
   }
 }
