@@ -1,16 +1,24 @@
 <script lang="ts">
-	export let isEnabled: boolean;
-	export let onChange: (value: boolean) => void;
+	import { run } from 'svelte/legacy';
 
-	let labelEl: HTMLLabelElement;
-	$: if (labelEl) {
-		labelEl.tabIndex = 0;
-		labelEl.onkeydown = (event) => {
-			if (event.target === event.currentTarget && (event.key === 'Enter' || event.key === ' ')) {
-				onChange(!isEnabled);
-			}
-		};
+	interface Props {
+		isEnabled: boolean;
+		onChange: (value: boolean) => void;
 	}
+
+	let { isEnabled, onChange }: Props = $props();
+
+	let labelEl: HTMLLabelElement = $state();
+	run(() => {
+		if (labelEl) {
+			labelEl.tabIndex = 0;
+			labelEl.onkeydown = (event) => {
+				if (event.target === event.currentTarget && (event.key === 'Enter' || event.key === ' ')) {
+					onChange(!isEnabled);
+				}
+			};
+		}
+	});
 </script>
 
 <label
@@ -18,5 +26,5 @@
 	class="checkbox-container cursor-pointer focus-visible:shadow-[0_0_0_3px_var(--background-modifier-border-focus)] outline-none"
 	class:is-enabled={isEnabled}
 >
-	<input type="checkbox" tabindex="-1" on:change={() => onChange(!isEnabled)} class="hidden" />
+	<input type="checkbox" tabindex="-1" onchange={() => onChange(!isEnabled)} class="hidden" />
 </label>
