@@ -1,16 +1,25 @@
 <script lang="ts">
 	import { LoadingCircle, settingsStore, type PeriodSettings } from '@/settings';
 	import { FilepathModal } from '@/ui/modals/filepath-select';
-	import clsx from 'clsx';
+	import { cn } from '@/ui/utils';
 
-	export let format: PeriodSettings['formats'][number];
-	export let selected: boolean = false;
-	export let error: string = '';
-	export let separator: string = ' • ';
+	interface Props {
+		format: PeriodSettings['formats'][number];
+		selected?: boolean;
+		error?: string;
+		separator?: string;
+	}
 
-	$: filepaths = $settingsStore.filepathsByFormatValue[format.value] || {};
-	$: filesCount = Object.keys(filepaths).length;
-	$: _selected = selected;
+	let {
+		format,
+		selected = false,
+		error = '',
+		separator = ' • '
+	}: Props = $props();
+
+	let filepaths = $derived($settingsStore.filepathsByFormatValue[format.value] || {});
+	let filesCount = $derived(Object.keys(filepaths).length);
+	let _selected = $derived(selected);
 
 	function handleShowFiles(filepaths: string[]) {
 		new FilepathModal(filepaths, format.value).open();
@@ -21,7 +30,7 @@
 	<div class="relative inline-flex">
 		<LoadingCircle loading={format.loading} />
 		<a
-			class={clsx(
+			class={cn(
 				'[font-size:calc(var(--font-ui-small)+1px)] focus-visible:shadow-[0_0_0_3px_var(--background-modifier-border-focus)]',
 				error
 					? 'text-[var(--text-error)] hover:opacity-85'
@@ -31,7 +40,7 @@
 				format.loading && 'opacity-60'
 			)}
 			href={'#'}
-			on:click={() => handleShowFiles(Object.keys(filepaths))}
+			onclick={() => handleShowFiles(Object.keys(filepaths))}
 			>{filesCount || '-'} {filesCount === 1 ? 'File' : 'Files'}</a
 		>
 	</div>
@@ -39,6 +48,6 @@
 {/if}
 
 <style lang="postcss">
+	@tailwind base;
 	@tailwind utilities;
 </style>
-
