@@ -10,9 +10,13 @@
 	import OpenAtStartup from './OpenAtStartup.svelte';
 	import Formats from './Formats.svelte';
 
-	export let granularity: IGranularity;
+	interface Props {
+		granularity: IGranularity;
+	}
 
-	let isExpanded = false;
+	let { granularity }: Props = $props();
+
+	let isExpanded = $state(false);
 
 	let settings = writableDerived(
 		settingsStore,
@@ -45,7 +49,7 @@
 	<a
 		href="./"
 		class="setting-item setting-item-heading text-transparent p-6 cursor-pointer flex items-center justify-between focus-visible:shadow-[0_0_0_3px_var(--background-modifier-border-focus)] outline-none"
-		on:click={toggleExpand}
+		onclick={toggleExpand}
 	>
 		<div class="setting-item-info flex justify-between items-center">
 			<h3 class="setting-item-name flex items-center text-lg font-semibold">
@@ -76,5 +80,6 @@
 
 <style lang="postcss">
 	@tailwind base;
+	@tailwind components;
 	@tailwind utilities;
 </style>
