@@ -1,6 +1,8 @@
 <svelte:options immutable />
 
 <script lang="ts">
+	import { run } from 'svelte/legacy';
+
 	import type { Moment } from 'moment';
 
 	import { activeFilepathStore, displayedDateStore } from '@/stores/';
@@ -9,42 +11,52 @@
 	import Sticker from './Sticker.svelte';
 	import { getFileData } from '@/io';
 	import { settingsStore } from '@/settings';
+	import { cn } from '@/ui/utils';
+	import { todayStore } from '@/stores/dates';
 
-	// Properties
-	export let date: Moment;
+	
+	interface Props {
+		// Properties
+		date: Moment;
+	}
 
-	// update today value in case the displayed date changes and the current date is no longer today
-	let today: Moment;
-	$: $displayedDateStore, (today = window.moment());
+	let { date }: Props = $props();
 
 	// file obtained with filename generated using selected format for given period
-	let { file, sticker } = getFileData('day', date);
-	$: {
+	let { file, sticker } = $state(getFileData('day', date));
+	run(() => {
 		if ($settingsStore.filepaths) {
 			const fileData = getFileData('day', date);
 			file = fileData.file;
 			sticker = fileData.sticker;
 		}
-	}
-	$: isActive = $activeFilepathStore === file?.path;
-	$: isToday = date.isSame(today, 'day');
-	$: isAdjacentMonth = !date.isSame($displayedDateStore, 'month');
+	});
+	let isActive = $derived($activeFilepathStore === file?.path);
+	let isToday = $derived(date.isSame($todayStore, 'day'));
+	let isAdjacentMonth = $derived(!date.isSame($displayedDateStore, 'month'));
 </script>
 
-<td class="relative">
+<td>
 	<button
-		class:active={isActive}
-		class:today={isToday}
-		class:adjacent-month={isAdjacentMonth}
-		class="[&:not(:focus-visible)]:shadow-none !h-auto w-full flex flex-col font-medium rounded-[--radius-s] text-sm px-1 py-3 relative text-center tabular-nums transition-colors day"
+		class={cn(
+			'relative [&:not(:focus-visible)]:shadow-none !h-auto w-full flex flex-col font-medium rounded-[--radius-s] text-sm px-1 pt-2.5 pb-4 text-center tabular-nums transition-colors',
+			isActive
+				? 'text-[--text-on-accent] bg-[--interactive-accent] hover:bg-[--interactive-accent-hover]'
+				: isToday
+				? 'text-[--color-text-today] hover:bg-[--interactive-hover]'
+				: isAdjacentMonth
+				? 'opacity-25 hover:bg-[--interactive-hover]'
+				: 'text-[--text-normal] hover:bg-[--interactive-hover]',
+			isActive && isToday && 'text-[--text-on-accent]'
+		)}
 		id="day"
-		on:click={(event) =>
+		onclick={(event) =>
 			eventHandlers.onClick({
 				date,
 				createNewSplitLeaf: isControlPressed(event),
 				granularity: 'day'
 			})}
-		on:contextmenu={(event) =>
+		oncontextmenu={(event) =>
 			eventHandlers.onContextMenu({
 				event,
 				fileData: {
@@ -54,7 +66,7 @@
 				date,
 				granularity: 'day'
 			})}
-		on:pointerenter={(event) => {
+		onpointerenter={(event) => {
 			eventHandlers.onHover({
 				targetEl: event.target,
 				isControlPressed: isControlPressed(event),
@@ -63,29 +75,9 @@
 		}}
 	>
 		{date.format('D')}
-		<Dot className="absolute bottom-1" isVisible={!!file} isFilled={!!file} isActive={!!file} />
+		<div class="absolute leading-[0] bottom-[calc(1rem/2)] translate-y-1/3">
+			<Dot isVisible={!!file} isFilled={!!file} {isActive} />
+		</div>
 	</button>
 	<Sticker sticker={sticker?.emoji} />
 </td>
-
-<style lang="postcss">
-	@tailwind base;
-	@tailwind components;
-	@tailwind utilities;
-
-	.day {
-		@apply text-[--text-normal] hover:bg-[--interactive-hover];
-	}
-	.day.active {
-		@apply text-[--text-on-accent] bg-[--interactive-accent] hover:bg-[--interactive-accent-hover];
-	}
-	.day.today {
-		@apply text-[--color-text-today];
-	}
-	.day.active.today {
-		@apply text-[--text-on-accent];
-	}
-	.day.adjacent-month {
-		@apply opacity-25;
-	}
-</style>
