@@ -1,36 +1,54 @@
 <svelte:options immutable />
 
 <script lang="ts">
+	import { run } from 'svelte/legacy';
+
 	import { getFileData } from '@/io';
-	import { displayedDateStore } from '@/stores';
-	import { eventHandlers, isControlPressed } from '../utils';
+	import { activeFilepathStore, displayedDateStore } from '@/stores';
+	import { cn, eventHandlers, isControlPressed } from '../utils';
 	import Dot from './Dot.svelte';
 	import Sticker from './Sticker.svelte';
 	import { settingsStore } from '@/settings';
+	import { todayStore } from '@/stores/dates';
 
-	export let year: number;
+	interface Props {
+		year: number;
+	}
 
-	$: date = $displayedDateStore.clone().year(year).startOf('year');
-	let { file, sticker } = getFileData('year', date);
-	$: {
+	let { year }: Props = $props();
+	let date = $derived($displayedDateStore.clone().year(year).startOf('year'));
+
+	let { file, sticker } = $state(getFileData('year', date));
+	run(() => {
 		if ($settingsStore.filepaths) {
 			const fileData = getFileData('year', date);
 			file = fileData.file;
 			sticker = fileData.sticker;
 		}
-	}
+	});
+	let isActive = $derived($activeFilepathStore === file?.path);
+	let isToday = $derived(date.isSame($todayStore, 'year'));
 </script>
 
-<td class="relative">
+<td>
 	<button
+		class={cn(
+			'relative [&:not(:focus-visible)]:shadow-none !h-auto w-full flex flex-col font-medium rounded-[--radius-s] text-sm px-1 pt-2.5 pb-4 text-center tabular-nums transition-colors',
+			isActive
+				? 'text-[--text-on-accent] bg-[--interactive-accent] hover:bg-[--interactive-accent-hover]'
+				: isToday
+				? 'text-[--color-text-today] hover:bg-[--interactive-hover]'
+				: 'text-[--text-normal] hover:bg-[--interactive-hover]',
+			isActive && isToday && 'text-[--text-on-accent]'
+		)}
 		id="year"
-		on:click={(event) =>
+		onclick={(event) =>
 			eventHandlers.onClick({
 				date,
 				createNewSplitLeaf: isControlPressed(event),
 				granularity: 'year'
 			})}
-		on:contextmenu={(event) =>
+		oncontextmenu={(event) =>
 			eventHandlers.onContextMenu({
 				event,
 				fileData: {
@@ -40,7 +58,7 @@
 				date,
 				granularity: 'year'
 			})}
-		on:pointerenter={(event) => {
+		onpointerenter={(event) => {
 			eventHandlers.onHover({
 				targetEl: event.target,
 				isControlPressed: isControlPressed(event),
@@ -49,7 +67,9 @@
 		}}
 	>
 		{year}
-		<Dot isFilled={!!file} isActive={!!file} />
+		<div class="absolute leading-[0] bottom-[calc(1rem/2)] translate-y-[35%]">
+			<Dot isVisible={!!file} isFilled={!!file} {isActive} />
+		</div>
 	</button>
 
 	<Sticker sticker={sticker?.emoji} />
@@ -57,6 +77,5 @@
 
 <style lang="postcss">
 	@tailwind base;
-	@tailwind components;
 	@tailwind utilities;
 </style>
