@@ -1,40 +1,50 @@
 <script lang="ts">
 	import { LoadingCircle } from '@/settings';
-	import clsx from 'clsx';
+	import { cn } from '@/ui/utils';
 	import { setIcon } from 'obsidian';
 	import { onMount } from 'svelte';
 
-	export let onClick: () => void;
-	export let ariaLabel: string = 'Delete note';
-	export let text: string = '';
-	export let loading: boolean = false;
+	interface Props {
+		onClick: () => void;
+		ariaLabel?: string;
+		text?: string;
+		loading?: boolean;
+	}
 
-	let deleteBttnEl: HTMLElement;
+	let {
+		onClick,
+		ariaLabel = 'Delete note',
+		text = '',
+		loading = false
+	}: Props = $props();
+
+	let deleteBttnEl: HTMLElement = $state();
 	onMount(() => {
 		setIcon(deleteBttnEl, 'lucide-trash-2');
 	});
 </script>
 
 <div
-	class={clsx(
+	class={cn(
 		'relative menu-item tappable is-warning text-xs flex items-center',
 		loading
 			? 'opacity-60 hover:bg-transparent cursor-not-allowed '
 			: 'hover:bg-[var(--background-modifier-hover)] cursor-pointer'
 	)}
 	aria-label={ariaLabel}
-	on:click={(e) => {
+	onclick={(e) => {
 		e.stopPropagation();
 		!loading && onClick();
 	}}
 >
 	<LoadingCircle {loading} />
-	<div class="menu-item-icon" bind:this={deleteBttnEl} />
+	<div class="menu-item-icon" bind:this={deleteBttnEl}></div>
 	{#if text}
 		<span>{text}</span>
 	{/if}
 </div>
 
 <style lang="postcss">
+	@tailwind base;
 	@tailwind utilities;
 </style>
