@@ -6,9 +6,10 @@ import { setupLocale } from "@/stores";
 import { genNoticeFragment } from "@/ui/utils";
 import { capitalize } from "@/utils";
 import { App, Notice, PluginSettingTab } from 'obsidian';
-import { SvelteComponent } from "svelte";
+import { SvelteComponent, mount } from "svelte";
 import { get } from "svelte/store";
 import { settingsStore } from "./store";
+import View from "@/View.svelte";
 
 export class SettingsTab extends PluginSettingTab {
     public plugin: PeriodicNotesCalendarPlugin;
@@ -24,15 +25,18 @@ export class SettingsTab extends PluginSettingTab {
     display() {
         this.containerEl.empty();
 
-        this.view = new Settings({
-            target: this.containerEl
-        })
+        this.view = mount(View, {
+                    target: this.containerEl,
+                })
+        // this.view = new Settings({
+        //     target: this.containerEl
+        // })
     }
     hide() {
         super.hide()
         this.view?.$destroy()
 
-        let errors = [];
+        const errors = [];
 
         const periods = get(settingsStore).periods
         for (const [g, periodSettings] of Object.entries(periods)) {
